# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My REST API",
        "description": "This is the OpenAPI documentation for my Laravel REST API project.",
        "contact": {
            "name": "Mantas",
            "email": "mantas@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/listings/{listingId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comments by listing",
                "operationId": "edcf98a4eaf2ccd99fcf3fa1e6c5b36e",
                "parameters": [
                    {
                        "name": "listingId",
                        "in": "path",
                        "description": "Listing ID to fetch comments for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of comments"
                    },
                    "404": {
                        "description": "Listing not found"
                    }
                }
            }
        },
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get all comments",
                "operationId": "2e2597016871fa5dd2bdfccf912edc41",
                "responses": {
                    "200": {
                        "description": "List of comments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "operationId": "c6ce2c222c8502aadd6a287105e49307",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "listing_id",
                                    "content"
                                ],
                                "properties": {
                                    "listing_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Great listing!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comment details",
                "operationId": "dfbfe62d6ea6efe5a6ebc2be766a00e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment details"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update an existing comment",
                "operationId": "17f16fb86b3722d33c54e6d165a02c80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Updated comment text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment",
                "operationId": "b683c30a239aa780ffc7e9b219e90811",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
...
    },
...
}